docs: handle POST request
serializer = ProductSerializer(data=request.data):
creates a new instance of ProductSerializer, passing in the incoming data
from the request (request.data). This prepares to validate and save this
data as a new product.

if serializer.is_valid():
calls the .is_valid() method on the serializer to check if the provided
data meets all validation criteria defined in the serializer.

instance = serializer.save():
If the data is valid, this line saves it to the database and returns the
newly created Product instance, only then the properties, instance
methods in the models will work etc.

return Response(serializer.data):
Returns a DRF Response, sending back the serialized data of the newly
created product as JSON.