feat: use generics.RetrieveAPIView for product details
https://www.django-rest-framework.org/api-guide/generic-views/#generic-views

https://www.django-rest-framework.org/api-guide/generic-views/#retrieveapiview

RetrieveAPIView is a class-based view provided by DRF that is specifically
designed for read-only endpoints. It is used to retrieve a single instance
of a model based on its primary key or another unique identifier.

- Read-Only: only supports HTTP GET method, making it suitable for fetching
  data without modifying it.

- Single Instance: Unlike ListAPIView, which returns a list of objects,
  RetrieveAPIView returns a single model instance

You can customize which field is used for lookup by setting the
lookup_field attribute. By default, this is set to "pk" (primary key),
but you can change it to any unique field in your model:

  class CustomPostDetail(RetrieveAPIView):
      queryset = Post.objects.all()
      serializer_class = PostSerializer
      lookup_field = 'slug'  # Use 'slug' instead of 'pk'